remove old tables
